2022 version/rl_course/requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
dm-control==1.0.7
dm-env==1.5
dm-tree==0.1.7
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
ffmpeg==1.4
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.5
gym==0.25.2
gym-notices==0.0.8
hydra-core==1.2.0
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipdb==0.13.9
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
labmaze==1.0.5
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
moviepy==1.0.3
mujoco==2.2.2
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
proglog==0.1.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
#pygame==2.1.2
Pygments==2.13.0
PyOpenGL==3.1.6
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
Send2Trash==1.8.0
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.5.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
urllib3==1.26.12
wandb==0.13.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.3
zipp==3.8.1